requirements.txt
torch
transformers
peft
bitsandbytes
trl
accelerate
datasets
openai
python-dotenv